requirements.txt
ultralytics
opencv-python
pygame
pyserial
numpy